+++ spi-impl/sml-file-config/src/main/java/org/n52/tasking/data/sml/device/DeviceParser.java
@@ -39,8 +39,8 @@ import org.n52.tasking.data.entity.Parameter;
 import org.n52.tasking.data.entity.QuantityParameter;
 import org.n52.tasking.data.entity.TaskingDescription;
 import org.n52.tasking.data.entity.TextParameter;
-import org.n52.tasking.data.sml.ParseException;
-import org.n52.tasking.data.sml.XPathParser;
+import org.n52.tasking.data.sml.xml.ParseException;
+import org.n52.tasking.data.sml.xml.XPathParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.w3c.dom.Node;
